+++ packages/server/src/routes/plot-census-routes.ts
@@ -1,7 +1,6 @@
-import { PlotCensus } from "@ong-forestry/schema";
 import express from "express";
 import { approve, getPlotCensuses, submitForReview } from "services";
-import { requireAuth } from "services/auth-service";
+import { requireAuth } from "middleware";
 
 const plotCensusRouter = express.Router();
