+++ packages/server/src/routes/tree-species-routes.ts
@@ -1,31 +1,34 @@
-import { TreeSpecies } from "@ong-forestry/schema";
-import { TreeSpeciesTypes } from "@ong-forestry/schema/src/tree";
 import express from "express";
-import { requireAuth } from "services/auth-service";
+import { TreeSpecies, TreeSpeciesTypes } from "@ong-forestry/schema";
 import { getTreeSpecies } from "services/tree-species-service";
+import { requireAuth } from "middleware";
 
-const treeSpeciesRouter = express.Router()
+const treeSpeciesRouter = express.Router();
 
 const parseParams = (query: any) => ({
-    code: query.code as string,
-    codes: query.codes as string[],
-    name: query.name as string,
-    family:query.family as string,
-    genus: query.genus as string,
-    commonName:query.commonName as string,
-    type:query.type as TreeSpeciesTypes,
-    limit: parseInt(query.limit as string),
-    offset: parseInt(query.offset as string),
-  });
+  code: query.code as string,
+  codes: query.codes as string[],
+  name: query.name as string,
+  family: query.family as string,
+  genus: query.genus as string,
+  commonName: query.commonName as string,
+  type: query.type as TreeSpeciesTypes,
+  limit: parseInt(query.limit as string),
+  offset: parseInt(query.offset as string),
+});
 
-treeSpeciesRouter.get<{}, any, TreeSpecies>('/', requireAuth, async(req,res)=>{
+treeSpeciesRouter.get<{}, any, TreeSpecies>(
+  "/",
+  requireAuth,
+  async (req, res) => {
     try {
-        const species = await getTreeSpecies(parseParams(req.query));
-        res.status(200).json(species);
-      } catch (e: any) {
-        console.error(e);
-        res.status(500).send(e?.message ?? "Unknown error.");
-      }
-})
+      const species = await getTreeSpecies(parseParams(req.query));
+      res.status(200).json(species);
+    } catch (e: any) {
+      console.error(e);
+      res.status(500).send(e?.message ?? "Unknown error.");
+    }
+  }
+);
 
-export {treeSpeciesRouter}
+export { treeSpeciesRouter };
